Move eventbridge models into shared folder, adding ReindexEvent
