+++ lambda/recipes-responder/tsconfig.exec.json
@@ -11,5 +11,5 @@
 		}
 	},
 	"exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"],
-	"include": ["src/**/*.ts"]
+	"include": ["src/**/*.ts", "../../lib/recipes-data/src/lib/eventbridge-models.ts"]
 }
